+++ src/compiler/scala/tools/nsc/interactive/tests/core/SourcesCollector.scala
@@ -13,7 +13,7 @@ private[tests] object SourcesCollector {
    * */
   def apply(base: Path, filter: SourceFilter): Array[SourceFile] = {
     assert(base.isDirectory)
-    base.walk.filter(filter).map(source).toArray
+    base.walk.filter(filter).map(source).toList.toArray.sortBy(_.file.name)
   }
 
   private def source(file: Path): SourceFile = source(AbstractFile.getFile(file.toFile))
